backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
azure-cosmos
